(refactor) move the "components" directory content to "shared".
