+++ src/app/app.component.spec.ts
@@ -4,7 +4,6 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { IgoModule } from 'igo2';
 
 import { SharedModule } from './shared/shared.module';
-import { ComponentModule } from './components/components.module';
 import { AppRoutingModule } from './app-routing';
 
 import { AppComponent } from './app.component';
@@ -17,7 +16,6 @@ describe('AppComponent', () => {
         IgoModule.forRoot(),
 
         SharedModule,
-        ComponentModule
       ],
       declarations: [
         AppComponent
