+++ src/app/shared/components/search-box/search-box.component.spec.ts
