+++ src/app/shared/components/search-list/search-list.component.ts
