+++ src/app/shared/interfaces/service-for-list.interface.ts
