--- src/app/shared/models/menu-item.interface.ts
@@ -1,5 +1,0 @@
-export interface MenuItem {
-  name: string;
-  title: string;
-  tooltip?: string;
-}
