+++ src/app/shared/pipes/filter.pipe.ts
